Currently, resizing is broken in FF9 and Chrome15, and is throwing errors. This is due to the fact that within the window resize event handler, "self.element.context" is set to window.document which does not contain the "clientWidth" and "clientHeight" properties, event hough these are being accessed widthin the window resize event handler.
2 possible alternatives to explore could be:

  1. Use `$(self.element.context).innerWidth()` and `$(self.element.context).innerHeight()`

  2. Use `self.element.context.body.clientWidth` and `self.element.context.body.clientHeight`

Unfortunately, while using these solutions eliminates the JS errors, the canvas is still not properly
redrawn. So for now, it seems that there is no problem with simply commenting
out the `self.resize()` call, which in my use cases will allow the canvas to remain as-is
without any rendering problems.